+++ packages/beta/src/types.ts
@@ -3,6 +3,7 @@
 import {
   Metadata,
   MetadataPatch,
+  SinglePatchRequired,
   SinglePatchRequiredString,
   SinglePatchString,
   Timestamp,
@@ -59,6 +60,7 @@ export const MonitoringTaskModelExternalId = {
 export interface MonitoringTaskThresholdModelCreateBase {
   externalId: MonitoringTaskModelExternalId;
 }
+
 export interface MonitoringTaskThresholdModelCreate
   extends MonitoringTaskThresholdModelCreateBase {
   externalId: 'threshold';
@@ -98,14 +100,18 @@ export interface MonitoringTaskCreate {
   externalId: CogniteExternalId;
   name: string;
   channelId: CogniteInternalId;
-  interval: number;
-  overlap: number;
+  interval?: number;
+  overlap?: number;
   model:
     | MonitoringTaskThresholdModelCreate
     | MonitoringTaskDoubleThresholdModelCreate;
   nonce: string;
   source?: string;
   sourceId?: string;
+  alertContext?: {
+    unsubscribeUrl: string;
+    investigateUrl: string;
+  };
 }
 
 export interface MonitoringTask {
@@ -224,12 +230,15 @@ export interface ChannelPatch {
     externalId?: SinglePatchString;
     name?: SinglePatchRequiredString;
     description?: SinglePatchString;
+    parentId?: SinglePatchRequired<number>;
     metadata?: MetadataPatch;
   };
 }
 
 export interface ChannelChangeById extends ChannelPatch, InternalId {}
+
 export interface ChannelChangeByExternalId extends ChannelPatch, ExternalId {}
+
 export type ChannelChange = ChannelChangeById | ChannelChangeByExternalId;
 
 export interface SubscriberCreate {