+++ packages/beta/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.5.1](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.5.0...@cognite/sdk-beta@5.5.1) (2023-11-28)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+# [5.5.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.4.3...@cognite/sdk-beta@5.5.0) (2023-11-28)
+
+### Features
+
+- monitoring tasks upsert ([#1042](https://github.com/cognitedata/cognite-sdk-js/issues/1042)) ([7e0e24c](https://github.com/cognitedata/cognite-sdk-js/commit/7e0e24cfb5c4d91302e54a270457b8d9b914ffd4))
+
+## [5.4.3](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.4.2...@cognite/sdk-beta@5.4.3) (2023-11-24)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+## [5.4.2](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.4.1...@cognite/sdk-beta@5.4.2) (2023-11-23)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+## [5.4.1](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.4.0...@cognite/sdk-beta@5.4.1) (2023-11-21)
+
+**Note:** Version bump only for package @cognite/sdk-beta
+
+# [5.4.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.3.0...@cognite/sdk-beta@5.4.0) (2023-11-01)
+
+### Features
+
+- **beta:** add alertContext to monitoring create type ([#1032](https://github.com/cognitedata/cognite-sdk-js/issues/1032)) ([ab21181](https://github.com/cognitedata/cognite-sdk-js/commit/ab211812f93a45fa2485a5d5464c30392d49f110))
+
 # [5.3.0](https://github.com/cognitedata/cognite-sdk-js/compare/@cognite/sdk-beta@5.2.34...@cognite/sdk-beta@5.3.0) (2023-10-17)
 
 ### Features